project is nearly done

+++ blog/app/Http/Controllers/AdminController.php
@@ -15,26 +15,33 @@ use App\Models\Post;
 use App\Models\Reply;
 use App\Models\Report;
 
+use Illuminate\Support\Facades\Gate;
+
 class AdminController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware("auth");
+    }
+
     //create category
-    public function category(){
-        $validator = validator(request()->all(),[
-            "category"=> "required"
+    public function category()
+    {
+        $validator = validator(request()->all(), [
+            "category" => "required"
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
         $category = new Category();
         $category->category = request()->category;
         $category->save();
         return back();
-
-
     }
 
-    public function categoryDelete($id){
+    public function categoryDelete($id)
+    {
         $data = Category::find($id);
         $data->delete();
         return back();
@@ -42,13 +49,14 @@ class AdminController extends Controller
 
 
     //advertisement section
-    public function adverAdd(){
-        $validator = validator(request()->all(),[
+    public function adverAdd()
+    {
+        $validator = validator(request()->all(), [
             "adver_id" => "required",
             "adver_img" => "required"
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
 
@@ -57,30 +65,32 @@ class AdminController extends Controller
 
         $imageName = time() . "." . $image->getClientOriginalExtension();
 
-        request()->adver_img->move("advertisementImg",$imageName);
+        request()->adver_img->move("advertisementImg", $imageName);
 
         $advertisement->adver_image = $imageName;
         $advertisement->save();
         return back();
     }
 
-    public function adverDelete($id){
+    public function adverDelete($id)
+    {
         $advertisement = Advertisement::find($id);
-        $advertisement->adver_image ="";
+        $advertisement->adver_image = "";
         $advertisement->save();
         return back();
     }
 
     //admin setting section
-    public function settingmanagement(){
-        $validator = validator(request()->all(),[
+    public function settingmanagement()
+    {
+        $validator = validator(request()->all(), [
             "key_data" => "required",
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
-        $key =request()->key_data;
+        $key = request()->key_data;
         $adminSetting = AdminSetting::find(1);
         $adminSetting->$key = request()->value_data ?? "";
         $adminSetting->save();
@@ -88,21 +98,23 @@ class AdminController extends Controller
     }
 
     //user controller section
-    public function userDel($id){
+    public function userDel($id)
+    {
         $user = User::find($id);
         $user->delete();
         return Back();
-
     }
 
-    public function userRole($id){
+    public function userRole($id)
+    {
         $user = User::find($id);
         $user->user_roll = "user";
         $user->save();
         return back();
     }
 
-    public function adminRole($id){
+    public function adminRole($id)
+    {
         $user = User::find($id);
         $user->user_roll = "admin";
         $user->save();
@@ -110,38 +122,42 @@ class AdminController extends Controller
     }
 
 
-    public function userBan($id){
+    public function userBan($id)
+    {
         $user = User::find($id);
         $user->ban = 1;
         $user->save();
         return back();
     }
 
-    public function userunBan($id){
+    public function userunBan($id)
+    {
         $user = User::find($id);
         $user->ban = 0;
         $user->save();
         return back();
     }
 
-    public function userSearch(){
-        $validator = validator(request()->all(),[
+    public function userSearch()
+    {
+        $validator = validator(request()->all(), [
             "search" => "required"
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
 
         $request_data = request()->search;
-        $user = User::where("name","like","%$request_data%")->orWhere("email","like","%$request_data%")->latest()->paginate(10);
-        return view("admin.siders.user",[
+        $user = User::where("name", "like", "%$request_data%")->orWhere("email", "like", "%$request_data%")->latest()->paginate(10);
+        return view("admin.siders.user", [
             "usersData" => $user
         ]);
     }
 
     //approve section
-    public function approve($id){
+    public function approve($id)
+    {
         $approve = Post::find($id);
         $approve->post_action = "approve";
         $approve->created_at  = Now();
@@ -149,11 +165,12 @@ class AdminController extends Controller
         return back();
     }
 
-    public function reject($id){
-        $validator = validator(request()->all(),[
+    public function reject($id)
+    {
+        $validator = validator(request()->all(), [
             "reject_type" => "required"
         ]);
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
         $post = Post::find($id);
@@ -163,9 +180,10 @@ class AdminController extends Controller
         return back();
     }
 
-    public function approveEverythings($counts){
-        $approve_request = Post::where("post_action","waiting")->get();
-        foreach($approve_request as $approve){
+    public function approveEverythings($counts)
+    {
+        $approve_request = Post::where("post_action", "waiting")->get();
+        foreach ($approve_request as $approve) {
             $approve->post_action = "approve";
             $approve->save();
         }
@@ -173,65 +191,84 @@ class AdminController extends Controller
     }
 
     //report management section
-    public function cancel($id){
+    public function cancel($id)
+    {
         $post = Post::find($id);
         $post->report = "none";
         $post->save();
 
-        $reports = Report::where("post_id",$id)->get();
-        foreach($reports as $report){
+        $reports = Report::where("post_id", $id)->get();
+        foreach ($reports as $report) {
             $report->delete();
         }
         return back();
     }
 
-    public function reportDelete($id){
+    public function reportDelete($id)
+    {
         $post = Post::find($id);
         $post->report = "delete";
         $post->save();
 
-        $reports = Report::where("post_id",$id)->get();
-        foreach($reports as $report){
+        $reports = Report::where("post_id", $id)->get();
+        foreach ($reports as $report) {
             $report->delete();
         }
         return back();
     }
 
-    public function reportShow($id){
+    public function reportShow($id)
+    {
         $post = Post::find($id);
-        return view("admin.parts.reportShow",[
+        return view("admin.parts.reportShow", [
             "postReports" => $post
         ]);
     }
 
     //comments check syste
-    public function checkComment(){
-        $validator = validator(request()->all(),[
+    public function checkComment()
+    {
+        $validator = validator(request()->all(), [
             "search" => "required"
         ]);
-        if($validator->fails()){
+        if ($validator->fails()) {
             return back()->withErrors($validator);
         }
 
         $request_data = request()->search;
-        $comments_data = Comment::where("content","like","%$request_data%")->get();
-        $replies_data  = Reply::where("content","like","%$request_data%")->get();
-        return view("admin.siders.comments",compact("comments_data","replies_data"));
+        $comments_data = Comment::where("content", "like", "%$request_data%")->get();
+        $replies_data  = Reply::where("content", "like", "%$request_data%")->get();
+        return view("admin.siders.comments", compact("comments_data", "replies_data"));
     }
 
-    public function deleteComment($id){
-        $comments_data = Comment::find($id);
-        $comments_data->delete();
+    public function deleteComment($id)
+    {
 
-        $replies_data = Reply::where("comment_id",$id)->get();
-        foreach($replies_data as $data){
-            $data->delete();
+        $user_id = Comment::find($id)->user_id;
+
+        if (Gate::allows("check-id", $user_id)) {
+            $comments_data = Comment::find($id);
+            $comments_data->delete();
+
+            $replies_data = Reply::where("comment_id", $id)->get();
+            foreach ($replies_data as $data) {
+                $data->delete();
+            }
+            return back();
+        } else {
+            return back();
         }
-        return back();
     }
-    public function deleteReply($id){
+
+    public function deleteReply($id)
+    {
+        $user_id = Reply::find($id)->user_id;
+        if (Gate::allows("check-id",$user_id )) {
         $replies_data = Reply::find($id);
         $replies_data->delete();
         return back();
+        }else{
+            return back();
+        }
     }
 }
